custom hooks for loading mockData
